refactor: bump auspice to 2.67.0
