+++ packages/nextclade-web/config/next/next.config.ts
@@ -106,7 +106,31 @@ const nextConfig: NextConfig = {
     styledComponents: true,
   },
   poweredByHeader: false,
-  transpilePackages: ['auspice', 'debug', 'jotai-devtools', 'semver'],
+  transpilePackages: [
+    'auspice',
+    'd3-array',
+    'd3-axis',
+    'd3-brush',
+    'd3-collection',
+    'd3-color',
+    'd3-dispatch',
+    'd3-drag',
+    'd3-ease',
+    'd3-format',
+    'd3-interpolate',
+    'd3-path',
+    'd3-scale',
+    'd3-selection',
+    'd3-shape',
+    'd3-time',
+    'd3-time-format',
+    'd3-timer',
+    'd3-transition',
+    'd3-zoom',
+    'debug',
+    'jotai-devtools',
+    'semver',
+  ],
   env,
 }
