+++ packages/nextclade-web/package.json
@@ -101,7 +101,7 @@
     "@tanstack/react-table": "8.21.3",
     "@tanstack/react-table-devtools": "8.21.3",
     "@use-gesture/react": "10.3.1",
-    "auspice": "2.59.1",
+    "auspice": "2.67.0",
     "axios": "1.9.0",
     "base64-js": "1.5.1",
     "bootstrap": "4.6.1",
