+++ packages/nextclade-web/src/state/reducer.ts
@@ -9,12 +9,12 @@ import browserDimensions from 'auspice/src/reducers/browserDimensions'
 import controls from 'auspice/src/reducers/controls'
 import entropy from 'auspice/src/reducers/entropy'
 import frequencies from 'auspice/src/reducers/frequencies'
-import measurements from 'auspice/src/reducers/measurements'
 import metadata from 'auspice/src/reducers/metadata'
 import narrative from 'auspice/src/reducers/narrative'
 import notifications from 'auspice/src/reducers/notifications'
 import tree from 'auspice/src/reducers/tree'
-import treeToo from 'auspice/src/reducers/treeToo'
+import treeToo from 'auspice/src/reducers/tree/treeToo'
+import measurements from 'auspice/src/reducers/measurements'
 // END reducers from auspice
 
 import { auspiceGeneralReducer, auspiceQueryReducer } from './auspice/auspice.reducer'
