+++ packages/nextclade-web/src/types/auspice/src/components/controls/toggle-focus.d.ts
@@ -1,10 +1,5 @@
 declare module 'auspice/src/components/controls/toggle-focus' {
-  import { FC, ReactElement } from 'react'
+  import { FC } from 'react'
 
-  export interface ToggleFocusProps {
-    tooltip: ReactElement
-  }
-
-  export const ToggleFocus: FC<ToggleFocusProps>
-  export default ToggleFocus
+  export const ToggleFocus: FC
 }
