+++ packages/nextclade-web/src/types/auspice/src/reducers/tree/index.d.ts
@@ -1,6 +1,6 @@
 declare module 'auspice/src/reducers/tree' {
   import { AuspiceTreeState } from 'auspice'
 
-  declare function tree(state?: TreeState): AuspiceTreeState | undefined
+  declare function tree(state?: AuspiceTreeState): AuspiceTreeState | undefined
   export default tree
 }
