+++ packages/nextclade-web/src/types/auspice/src/reducers/tree/treeToo.d.ts
@@ -1,4 +1,4 @@
-declare module 'auspice/src/reducers/treeToo' {
+declare module 'auspice/src/reducers/tree/treeToo' {
   import { AuspiceTreeState } from 'auspice'
 
   declare function treeToo(state?: AuspiceTreeState): AuspiceTreeState | undefined
